+++ EasyDriver.Tests/GlobalUsings.cs
@@ -1,0 +1,4 @@
+global using Xunit;
+global using static Comfast.EasyDriver.DriverApi;
+global using Comfast.EasyDriver;
+global using static EasyDriver.Tests.Util.Assertions;