+++ EasyDriver.Tests/Integration/AssertionsTest.cs
@@ -1,0 +1,10 @@
+﻿using FluentAssertions;
+
+namespace EasyDriver.Tests.Integration;
+
+public class AssertionsTest {
+    [Fact] void WaitForFail() {
+        ShouldThrow(() => S("html >> lolz").WaitFor(100),
+            "html >> lolz");
+    }
+}