+++ EasyDriver.Tests/UnitTest1.cs
@@ -1,6 +1,0 @@
-namespace EasyDriver.Tests;
-
-public class UnitTest1 {
-    [Fact]
-    public void Test1() { }
-}